Extract steps subsystem from dashboard-main.js
